Split MotionDataStore from MiscDataStore
